feat(pos): move scan-outage feedback from a standing banner to scan-time toasts
The red "Scanning unavailable — sync engine isn't ready" banner flashed
above the product grid on every app load (the bootstrap lifecycle gate),
alarming cashiers about a state in which scanning actually works: local
lookups never needed the engine, and gated engines still resolve unknown
codes online (see the use-barcode gated-engine spec). Engine outages now
surface only when a scan actually hits them:

- The offline scan toast carries a "View status" action linking to
  /health/database, replacing the banner's link, and its copy no longer
  points at a banner that doesn't exist.
- Opening the camera scanner while sync is starting, offline, or the
  local database is degraded shows a quiet note in the panel's status
  overlay instead of a red banner.

The banner itself survives only for the dead-storage-worker outage
(#163): that blocks scanning, checkout, saving and voiding at once, its
only recovery is a reload, and it never flashes at startup — renamed to
StorageOutageBanner with the engine branches removed.
