+++ packages/core/src/contexts/translations/locales/en/core.json
@@ -1007,10 +1007,8 @@
 	"pos_products.scan_lookup_failed": "Lookup failed",
 	"pos_products.scan_lookup_failed_description": "{code} — Store didn’t respond — check your connection and scan again",
 	"pos_products.scan_unavailable": "Scanning unavailable",
-	"pos_products.scan_unavailable_description": "{code} — Sync engine is offline — see the banner above the products",
+	"pos_products.scan_unavailable_description": "{code} — Not on this device, and the store can’t be searched while offline",
 	"pos_products.scan_storage_unavailable_description": "{code} — Local database unavailable — reload the app before taking more sales",
-	"pos_products.scan_outage_banner": "Scanning unavailable — sync engine offline.",
-	"pos_products.scan_engine_unavailable_banner": "Scanning unavailable — sync engine isn’t ready.",
 	"pos_products.scan_storage_outage_banner": "Local database unavailable — scanning, checkout, saving and voiding are blocked until the app is reloaded.",
 	"pos_products.scan_storage_outage_reload": "Reload the app",
 	"pos_products.scan_storage_outage_restart_manually": "This build can't reload itself — close the app completely and open it again.",
@@ -1032,6 +1030,9 @@
 	"pos_products.camera_scan_open": "Scan barcode with camera",
 	"pos_products.camera_starting": "Starting camera…",
 	"pos_products.camera_unavailable": "Camera unavailable — check that no other app is using it.",
+	"pos_products.camera_scan_storage_unavailable": "Local database unavailable — reload the app to restore scanning.",
+	"pos_products.camera_scan_offline": "Offline — only products already on this device will scan.",
+	"pos_products.camera_scan_engine_not_ready": "Sync is still starting — some products may not be found yet.",
 	"settings.scanner_connect_serial": "Connect serial scanner",
 	"settings.scanner_connect_hid": "Connect HID scanner",
 	"settings.scanner_serial_disconnect": "Disconnect serial scanner",
